requirements.txt
dapr==1.14.0
fastapi==0.109.1
grpcio>=1.65.4
pydantic==2.4.2
requests==2.31.0
dapr-ext-workflow>=0.5.0
yaspin==3.0.2
uvicorn==0.30.6